agentic project/hospital-agent/requirement.txt
agno>=1.0
fastapi
uvicorn[standard]
pandas
python-dotenv
requests
beautifulsoup4
